requirements.txt
absl-py @ file:///opt/conda/conda-bld/absl-py_1639803114343/work
aiohttp @ file:///C:/ci/aiohttp_1646806572557/work
aiosignal @ file:///tmp/build/80754af9/aiosignal_1637843061372/work
anyascii==0.3.1
astor==0.8.1
astunparse==1.6.3
async-timeout @ file:///tmp/build/80754af9/async-timeout_1637851218186/work
attrs @ file:///opt/conda/conda-bld/attrs_1642510447205/work
blinker==1.4
Bottleneck @ file:///C:/ci/bottleneck_1648010904582/work
brotlipy==0.7.0
cachetools @ file:///tmp/build/80754af9/cachetools_1619597386817/work
certifi==2021.10.8
cffi @ file:///C:/ci_310/cffi_1642682485096/work
charset-normalizer @ file:///tmp/build/80754af9/charset-normalizer_1630003229654/work
click @ file:///C:/ci/click_1646038595831/work
colorama @ file:///tmp/build/80754af9/colorama_1607707115595/work
contractions==0.1.72
cryptography @ file:///C:/ci/cryptography_1633520531101/work
cycler==0.11.0
Cython==0.29.28
dill==0.3.4
dm-tree==0.1.7
emoji==1.7.0
Flask @ file:///opt/conda/conda-bld/flask_1648041541647/work
flatbuffers==1.12
fonttools==4.33.3
frozenlist @ file:///C:/ci/frozenlist_1637767271796/work
gast @ file:///Users/ktietz/demo/mc3/conda-bld/gast_1628588903283/work
gin-config==0.5.0
google-api-core==2.7.3
google-api-python-client==2.47.0
google-auth==1.35.0
google-auth-httplib2==0.1.0
google-auth-oauthlib==0.4.1
google-pasta @ file:///Users/ktietz/demo/mc3/conda-bld/google-pasta_1630577991354/work
googleapis-common-protos==1.56.1
grpcio==1.34.1
h5py==3.1.0
httplib2==0.20.4
idna @ file:///tmp/build/80754af9/idna_1637925883363/work
importlib-metadata @ file:///C:/ci/importlib-metadata_1648562621412/work
itsdangerous @ file:///tmp/build/80754af9/itsdangerous_1621432558163/work
Jinja2 @ file:///opt/conda/conda-bld/jinja2_1647436528585/work
joblib @ file:///tmp/build/80754af9/joblib_1635411271373/work
kaggle==1.5.12
keras==2.8.0
keras-nightly==2.5.0.dev2021032900
Keras-Preprocessing @ file:///tmp/build/80754af9/keras-preprocessing_1612283640596/work
kiwisolver==1.4.2
libclang==14.0.1
Markdown @ file:///C:/ci/markdown_1614364082838/work
MarkupSafe @ file:///C:/ci/markupsafe_1621528502553/work
matplotlib==3.5.2
mkl-fft==1.3.1
mkl-random @ file:///C:/ci/mkl_random_1626186184308/work
mkl-service==2.4.0
multidict @ file:///C:/ci/multidict_1607349747897/work
nltk @ file:///opt/conda/conda-bld/nltk_1645628263994/work
numexpr @ file:///C:/ci/numexpr_1640704337920/work
numpy==1.22.3
oauth2client==4.1.3
oauthlib @ file:///opt/conda/conda-bld/oauthlib_1644332107998/work
opencv-python-headless==4.5.5.64
opt-einsum @ file:///tmp/build/80754af9/opt_einsum_1621500238896/work
packaging @ file:///tmp/build/80754af9/packaging_1637314298585/work
pandas @ file:///C:/ci/pandas_1650373458095/work
Pillow==9.1.0
portalocker==2.4.0
promise==2.3
protobuf==3.20.1
psutil==5.9.0
py-cpuinfo==8.0.0
pyahocorasick==1.4.4
pyasn1 @ file:///Users/ktietz/demo/mc3/conda-bld/pyasn1_1629708007385/work
pyasn1-modules==0.2.8
pycocotools==2.0.4
pycparser @ file:///tmp/build/80754af9/pycparser_1636541352034/work
PyJWT @ file:///C:/ci/pyjwt_1619682721924/work
pyOpenSSL @ file:///tmp/build/80754af9/pyopenssl_1635333100036/work
pyparsing @ file:///tmp/build/80754af9/pyparsing_1635766073266/work
pyreadline==2.1
PySocks @ file:///C:/ci/pysocks_1605307512533/work
python-dateutil @ file:///tmp/build/80754af9/python-dateutil_1626374649649/work
python-slugify==6.1.2
pytz==2021.3
pywin32==304
PyYAML==5.4.1
regex @ file:///C:/ci/regex_1648447888413/work
requests @ file:///opt/conda/conda-bld/requests_1641824580448/work
requests-oauthlib==1.3.0
rsa @ file:///tmp/build/80754af9/rsa_1614366226499/work
sacrebleu==2.0.0
scikit-learn==1.1.0
scipy @ file:///C:/ci/scipy_1641555170412/work
sentencepiece==0.1.96
seqeval==1.2.2
six==1.15.0
tabulate==0.8.9
tensorboard==2.8.0
tensorboard-data-server @ file:///C:/ci/tensorboard-data-server_1633035225228/work/tensorboard_data_server-0.6.0-py3-none-any.whl
tensorboard-plugin-wit==1.6.0
tensorflow==2.8.0
tensorflow-addons==0.16.1
tensorflow-datasets==4.5.2
tensorflow-estimator==2.5.0
tensorflow-hub==0.12.0
tensorflow-io-gcs-filesystem==0.25.0
tensorflow-metadata==1.8.0
tensorflow-model-optimization==0.7.2
tensorflow-text==2.8.2
termcolor==1.1.0
text-unidecode==1.3
textsearch==0.0.21
tf-estimator-nightly==2.8.0.dev2021122109
tf-models-official==2.8.0
tf-slim==1.1.0
threadpoolctl==3.1.0
tqdm @ file:///C:/ci/tqdm_1650636210717/work
typeguard==2.13.3
typing-extensions==3.7.4.3
uritemplate==4.1.1
urllib3 @ file:///C:/ci/urllib3_1650639883891/work
Werkzeug @ file:///opt/conda/conda-bld/werkzeug_1645628268370/work
win-inet-pton @ file:///C:/ci/win_inet_pton_1605306162074/work
wincertstore==0.2
wrapt==1.12.1
yarl @ file:///C:/ci/yarl_1606940155993/work
zipp @ file:///C:/ci/zipp_1652273994994/work
